+++ src/routes/cart/-components/index.ts
@@ -1,2 +1,4 @@
 export * from './DeleteConfirmModal';
 export * from './OrderConfirmModal';
+export * from './CartItemList';
+export * from './CartOrderInfo';
