+++ src/routes/order/-components/index.ts
@@ -1,0 +1,2 @@
+export * from './OrderItemList';
+export * from './OrderTotal';
