+++ src/routes/products/-components/index.ts
@@ -1,0 +1,2 @@
+export * from './ProductCard';
+export * from './ProductList';
